requirements.txt
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
python-telegram-bot==20.7
beautifulsoup4==4.12.2
selenium==4.17.2
webdriver-manager==4.0.1
pytz==2024.1